Moved FXServiceViewActor to "controller.actors" package
