+++ client/src/main/scala/it/cwmp/client/controller/actors/RoomViewActor.scala
@@ -4,7 +4,6 @@ import it.cwmp.client.controller.actors.RoomViewActor._
 import it.cwmp.client.controller.messages.AuthenticationRequests.GUILogOut
 import it.cwmp.client.controller.messages.GUIRequest
 import it.cwmp.client.controller.messages.RoomsRequests._
-import it.cwmp.client.view.FXServiceViewActor
 import it.cwmp.client.view.room.{RoomFXController, RoomStrategy}
 
 /**
